refactor: reorganize DTOs into project/section/task packages and update imports
