--- app/src/main/java/com/example/twdist_android/features/projectdetails/data/dto/SectionResponseDto.kt
@@ -1,12 +1,0 @@
-package com.example.twdist_android.features.projectdetails.data.dto
-
-import com.example.twdist_android.core.network.FlexibleStringIdSerializer
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class SectionResponseDto(
-    @Serializable(with = FlexibleStringIdSerializer::class)
-    val id: String,
-    val name: String,
-    val taskIds: List<String> = emptyList()
-)
