+++ app/src/main/java/com/example/twdist_android/features/projectdetails/data/dto/section/SectionResponseDto.kt
@@ -1,13 +1,12 @@
-package com.example.twdist_android.features.projectdetails.data.dto
+package com.example.twdist_android.features.projectdetails.data.dto.section
 
 import com.example.twdist_android.core.network.FlexibleStringIdSerializer
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ProjectDetailResponseDto(
+data class SectionResponseDto(
     @Serializable(with = FlexibleStringIdSerializer::class)
     val id: String,
     val name: String,
-    val favorite: Boolean,
-    val sections: List<SectionResponseDto> = emptyList()
+    val taskIds: List<String> = emptyList()
 )
