+++ app/src/main/java/com/example/twdist_android/features/projectdetails/data/dto/task/TaskResponseDto.kt
@@ -1,4 +1,4 @@
-package com.example.twdist_android.features.projectdetails.data.dto
+package com.example.twdist_android.features.projectdetails.data.dto.task
 
 import kotlinx.serialization.Serializable
