+++ app/src/main/java/com/example/twdist_android/features/projectdetails/data/mapper/TaskCompletionMapper.kt
@@ -1,6 +1,6 @@
 package com.example.twdist_android.features.projectdetails.data.mapper
 
-import com.example.twdist_android.features.projectdetails.data.dto.CompleteTaskRequestDto
+import com.example.twdist_android.features.projectdetails.data.dto.task.CompleteTaskRequestDto
 import java.time.LocalDate
 
 fun LocalDate?.toCompleteTaskRequestDto(): CompleteTaskRequestDto {
