+++ app/src/main/java/com/example/twdist_android/features/projectdetails/data/repository/TaskRepositoryImpl.kt
@@ -2,8 +2,8 @@ package com.example.twdist_android.features.projectdetails.data.repository
 
 import com.example.twdist_android.core.coroutines.runSuspendCatching
 import com.example.twdist_android.features.projectdetails.data.mapper.toCompleteTaskRequestDto
-import com.example.twdist_android.features.projectdetails.data.dto.CreateTaskRequestDto
-import com.example.twdist_android.features.projectdetails.data.dto.UpdateTaskRequestDto
+import com.example.twdist_android.features.projectdetails.data.dto.task.CreateTaskRequestDto
+import com.example.twdist_android.features.projectdetails.data.dto.task.UpdateTaskRequestDto
 import com.example.twdist_android.features.projectdetails.data.mapper.toDomainTask
 import com.example.twdist_android.features.projectdetails.data.remote.ProjectDetailsApi
 import com.example.twdist_android.features.projectdetails.domain.model.Task
