+++ app/src/test/java/com/example/twdist_android/features/projectdetails/data/mapper/TaskMapperTest.kt
@@ -1,6 +1,6 @@
 package com.example.twdist_android.features.projectdetails.data.mapper
 
-import com.example.twdist_android.features.projectdetails.data.dto.TaskResponseDto
+import com.example.twdist_android.features.projectdetails.data.dto.task.TaskResponseDto
 import org.junit.Assert.assertEquals
 import org.junit.Test
 import java.time.Instant
